style: 띄어쓰기 및 코드 정리
