+++ src/views/Home/components/MiddleSection.vue
